requirements.txt
# Core
Django==5.2.2

# REST API
djangorestframework
drf-nested-routers
drf-spectacular
drf-spectacular-sidecar

# JWT Authentication
djangorestframework-simplejwt

# Environment
django-environ==0.12.0

# HTTP Client
requests==2.31.0

# Code quality
flake8==7.3.0
